+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.1.0-rc.17",
+  sdkVersion: "4.1.0-rc.19",
   genVersion: "2.762.0",
-  userAgent: "speakeasy-sdk/typescript 4.1.0-rc.17 2.762.0 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 4.1.0-rc.19 2.762.0 2.0 @orq-ai/node",
 } as const;