+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -608,7 +608,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KB4E5QNP0703RFEKJ0EAFZDG"),
+  _id: z.string().default("01KB56975Z6NEQ9EGRCKEDGQFA"),
   display_name: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$inboundSchema,
@@ -651,7 +651,7 @@ export const CreateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01KB4E5QNP0703RFEKJ0EAFZDG"),
+  id: z.string().default("01KB56975Z6NEQ9EGRCKEDGQFA"),
   displayName: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$outboundSchema,