+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -211,7 +211,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-11-28T05:14:49.673Z",
+    "2025-11-28T12:16:10.264Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -250,7 +250,7 @@ export const ListDatasetsData$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-11-28T05:14:49.673Z"))
+  updated: z.date().default(() => new Date("2025-11-28T12:16:10.264Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {