+++ packages/orq-rc/RELEASES.md
@@ -5459,3 +5459,13 @@ Based on:
 - [typescript v4.1.0-rc.17] packages/orq-rc
 ### Releases
 - [NPM v4.1.0-rc.17] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.17 - packages/orq-rc
+
+## 2025-11-28 12:18:01
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.661.0 (2.762.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.1.0-rc.19] packages/orq-rc
+### Releases
+- [NPM v4.1.0-rc.19] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.19 - packages/orq-rc